refactor: ViewController와 UserViewController
GetMapping은 UserViewController에
